Presenter now simply uses RetainedFragment as a second view.
Updated so that the observable doesn't need to be passed around between RetainedFragment, Activity and Presenter.

It will get set directly between the Presenter and "RetainedView".
So it will just be as if this Presenter has two Views.
This is done simply by setting the RetainedFragment as part of the Presenter's constructor.